Image Customizer: Move partition utils
